# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenEAAP API
  description: |
    OpenEAAP (Open Enterprise AI Agent Platform) REST API Specification.

    This API provides endpoints for managing AI agents, workflows, models, and knowledge bases.

    ## Authentication
    All API requests require authentication using Bearer tokens:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Rate Limiting
    - Default: 100 requests per minute per user
    - Burst: 200 requests per minute

    ## Versioning
    API version is included in the path: `/api/v1/...`
  version: 1.0.0
  contact:
    name: OpenEAAP Team
    email: support@openeeap.org
    url: https://github.com/openeeap/openeeap
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.openeeap.org/v1
    description: Production server
  - url: https://staging-api.openeeap.org/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /agents:
    get:
      tags: [agents]
      summary: List all agents
      description: Retrieve a paginated list of agents
      operationId: listAgents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: runtime_type
          in: query
          schema:
            type: string
            enum: [native, langchain, autogpt]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []

    post:
      tags: [agents]
      summary: Create a new agent
      description: Create and deploy a new AI agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /agents/{agentId}:
    get:
      tags: [agents]
      summary: Get agent by ID
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    put:
      tags: [agents]
      summary: Update agent
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    delete:
      tags: [agents]
      summary: Delete agent
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '204':
          description: Agent deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
  /agents/{agentId}/execute:
    post:
      tags: [agents]
      summary: Execute agent
      description: Execute an agent with the provided input
      operationId: executeAgent
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteAgentRequest'
      responses:
        '200':
          description: Execution completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    AgentIdParam:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        runtime_type:
          type: string
          enum: [native, langchain, autogpt]
        config:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [active, inactive, error]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required:
        - name
        - runtime_type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        runtime_type:
          type: string
          enum: [native, langchain, autogpt]
        config:
          type: object
          additionalProperties: true
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        config:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [active, inactive]
    ExecuteAgentRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: string
        context:
          type: object
          additionalProperties: true
        parameters:
          type: object
          additionalProperties: true
    ExecutionResult:
      type: object
      properties:
        output:
          type: string
        metadata:
          type: object
          additionalProperties: true
        trace_id:
          type: string
        execution_time_ms:
          type: integer

    # ==================== Workflow Schemas ====================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# Personal.AI order the ending